Conflicts:
	static/build/css/main.css.map
	static/js/controllers/SearchResultCtrl.js
	templates/_media.html

+++ static/js/directives/KitinEntity.js
@@ -40,10 +40,10 @@ kitin.directive('kitinEntity', function(editService, $rootScope, $parse) {
     },
     template:   '<div class="{{classNames}}">' +
                   '<div ng-if="objects" ng-repeat="object in objects track by $index" class="{{innerClassNames}}"" ng-class="{auth: isLinked(object)}">' +
-                    '<span class="inner" ng-include="viewTemplate"></span>' +
+                    '<span class="inner" ng-include="viewTemplate()"></span>' +
                     '<span class="controls"><a class="delete" data-ng-click="doRemove($index)"><i class="fa fa-times"></i></a></span>' +
                   '</div>' +
-                  '<span ng-transclude></span>' +
+                  '<span ng-transclude ng-hide="hideAddControl()"></span>' +
                 '</div>',
 
     controller: function($element, $scope, $attrs) {
@@ -52,7 +52,14 @@ kitin.directive('kitinEntity', function(editService, $rootScope, $parse) {
         angular.extend($attrs, $scope.attributes);
       }
       
-      $scope.viewTemplate = $attrs.view || '/snippets/view-generic-linked-entity';
+      $scope.viewTemplate = function() {
+        var type = getType();
+        var view = $attrs.view || '/snippets/view-generic-linked-entity';
+        if ( type ) {
+          view = _.template(view)({ type: getType() });
+        }
+        return view;
+      };
      
       $scope.searchTemplate = $attrs.search;
 
@@ -82,6 +89,10 @@ kitin.directive('kitinEntity', function(editService, $rootScope, $parse) {
         $scope.objects = null;
       }
 
+      $scope.hideAddControl = function() {
+        return ($scope.objects && $scope.objects.length) && !$scope.multiple;
+      };
+
       var classNames = ['entity'];
       var innerClassNames = ['entity-content'];
       if ( $attrs.hasOwnProperty('rich') && $attrs.rich !== false) {
@@ -97,6 +108,31 @@ kitin.directive('kitinEntity', function(editService, $rootScope, $parse) {
       $scope.classNames = classNames.join(' ');
       $scope.innerClassNames = innerClassNames.join(' ');
 
+      var typeIndex = 0;
+      var types = [$attrs.type];
+
+      // allow multiple types
+      if ( /^\[.+\]$/.test($attrs.type) ) {
+        types = $scope.$eval($attrs.type);
+      }
+
+
+
+      // method for setting type if multiple types are supported
+      this.setType = function(index) {
+        typeIndex = index;
+      };
+
+      this.getTypes = function() {
+        return types;
+      };
+
+      function getType() {
+        return types[typeIndex];
+      }
+
+      this.getType = getType;
+
       this.doAdd = function(data) {
         var added = editService.addObject(subj, $scope.link, $scope.property, $scope.multiple, data);
 
@@ -110,29 +146,25 @@ kitin.directive('kitinEntity', function(editService, $rootScope, $parse) {
       };
 
       this.doCreate = function(initialValue) {
-        var type = $attrs.type;
-        // For subjects, creation is in an ng-repeat. Then try to eval variable to get type value from scope attribute
-        try { type = $scope.$eval($attrs.type); } catch(error) {}
-        if(_.isUndefined(type)) {
-          type = $attrs.type;
-        }
-        return editService.createObject($scope.property, type, initialValue);
+        return editService.createObject($scope.property, this.getType(), initialValue);
       };
 
       $scope.doRemove = function (index) {
-        var removed = null;
-        if ($scope.multiple && _.isNumber(index)) {
-          removed = subj[$scope.link].splice(index, 1)[0];
-        } else {
-          removed = subj[$scope.link];
-          delete subj[$scope.link];
-          $scope.objects = null;
-          $scope.viewmode = false;
+        if ( window.confirm('Är du säker på att du vill ta bort?') ) {
+          var removed = null;
+          if ($scope.multiple && _.isNumber(index)) {
+            removed = subj[$scope.link].splice(index, 1)[0];
+          } else {
+            removed = subj[$scope.link];
+            delete subj[$scope.link];
+            $scope.objects = null;
+            $scope.viewmode = false;
+          }
+          if (typeof subj.onRemove === 'function') {
+            subj.onRemove($scope.link, removed, index);
+          }
+          $scope.$emit('entity', $scope.objects);
         }
-        if (typeof subj.onRemove === 'function') {
-          subj.onRemove($scope.link, removed, index);
-        }
-        $scope.$emit('entity', $scope.objects);
       };
 
     }
